+++ lib/processors/ios/xcconfig/ios_xcconfig_processor.dart
@@ -23,16 +23,24 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
+import 'dart:collection';
+
+import 'package:flutter_flavorizr/extensions/extensions+map.dart';
 import 'package:flutter_flavorizr/parser/models/flavorizr.dart';
+import 'package:flutter_flavorizr/parser/models/flavors/flavor.dart';
+import 'package:flutter_flavorizr/parser/models/flavors/ios/enums.dart';
+import 'package:flutter_flavorizr/parser/models/flavors/ios/variable.dart';
 import 'package:flutter_flavorizr/processors/commons/string_processor.dart';
 
 class IOSXCConfigProcessor extends StringProcessor {
-  final String _appName;
   final String _flavorName;
+  final Flavor _flavor;
+  final Target? _target;
 
   IOSXCConfigProcessor(
-    this._appName,
-    this._flavorName, {
+    this._flavorName,
+    this._flavor,
+    this._target, {
     String? input,
     required Flavorizr config,
   }) : super(
@@ -55,11 +63,20 @@ class IOSXCConfigProcessor extends StringProcessor {
   }
 
   void _appendBody(StringBuffer buffer) {
+    final Map<String, Variable> variables = LinkedHashMap.from({
+      'FLUTTER_TARGET': Variable(value: 'lib/main-$_flavorName.dart'),
+      'ASSET_PREFIX': Variable(value: _flavorName),
+      'BUNDLE_NAME': Variable(value: _flavor.app.name),
+    })
+      ..addAll(
+        _flavor.ios.variables.where((_, variable) =>
+            variable.target == null || variable.target == _target),
+      );
+
     buffer.writeln();
-    buffer.writeln('FLUTTER_TARGET=lib/main-$_flavorName.dart');
-    buffer.writeln();
-    buffer.writeln('ASSET_PREFIX=$_flavorName');
-    buffer.writeln('BUNDLE_NAME=$_appName');
+    variables.forEach((key, variable) {
+      buffer.writeln('$key=${variable.value}');
+    });
   }
 
   @override